+++ test/e2e/global-setup.js
@@ -1,0 +1,10 @@
+import { execSync } from "node:child_process";
+
+export default function () {
+	execSync(
+		"npm run build:simple && npm run build:reactive && npm run build:gom",
+		{
+			stdio: "inherit",
+		},
+	);
+}